% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortof-package.R
\docType{package}
\name{sortof-package}
\alias{sortof}
\alias{sortof-package}
\title{sortof: Sort Stuff}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@rstudio.com}

}
\keyword{internal}
